Rename root package
