+++ beachfront/__init__.py
@@ -11,4 +11,15 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-from bfapi.service import piazza, algorithms, geoserver, scenes, jobs, productlines  # Order matters here
+import os
+
+import beachfront.logger
+
+DEBUG_MODE = os.getenv('DEBUG_MODE') == '1'
+MUTE_LOGS = os.getenv('MUTE_LOGS') == '1'
+
+
+beachfront.logger.init(
+    debug=DEBUG_MODE,
+    muted=MUTE_LOGS,
+)
