+++ beachfront/db/scenes.py
@@ -14,7 +14,7 @@
 import json
 from datetime import datetime
 import logging
-from bfapi.db import Connection
+from beachfront.db import Connection
 
 
 def insert(
